tests/helpers: Move pickUnusedPort, launchRedis
These helpers might make sense to move to a separate package one day.
